backend/requirements.txt
attrs==21.4.0
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-SocketIO==5.1.1
h11==0.13.0
idna==3.3
importlib-metadata==4.11.2
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
pluggy==1.0.0
propelauth-flask==1.1.0
propelauth-py==2.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.2
requests==2.27.1
simple-websocket==0.5.2
six==1.16.0
SQLAlchemy==1.4.32
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.1.0
zipp==3.7.0
